

\name{dbglm-inl}
\alias{dbglm-internal}
\alias{control_family}
\alias{controls_dbglm}
\alias{dbglm.iteration}

\title{Internal  functions}
\description{Internal  functions}
\usage{
control_family(family)
controls_dbglm(distance,weights,offset,rel.gvar,maxiter,eps1,eps2,y, method)
dbglm.iteration(y, mu, weights, nobs, eta, Delta, method, offset, n, eff.rank = NULL,
	rel.gvar, dev.resids, aic, mu.eta, valideta, validmu, family, variance, 
	linkinv, problem.links, eps1,eps2,maxiter)
}

\details{Not to be called by users}
\keyword{internal}
