
\name{dblm-inl}
\alias{dblm-internal}
\alias{Gcalc}
\alias{HwProject}
\alias{Hwyhat}
\alias{control_method}
\alias{control_metric}
\alias{controls_dblm}
\alias{formula_to_zy}

\title{Internal  functions}
\description{Internal  functions}
\usage{
Gcalc(n,weights,Delta)
HwProject(G,Dsqw,rk,epsilon,cvyes=FALSE,n)
Hwyhat(G,n,Dsqw,weights,rk,epsilon,y,y0,cvyes=FALSE,ori_weights,...)
control_method(method,class_call)
control_metric(metric)
controls_dblm(G,weights,eff.rank,rel.gvar,method,y)
formula_to_zy(formula,data,mf,class_mod,metric)
}

\details{Not to be called by users}
\keyword{internal}
