
\name{ldbglm-inl}
\alias{ldbglm-internal}
\alias{pred.train.sample.dbglm}


\title{Internal  functions}
\description{Internal  functions}
\usage{
pred.train.sample.dbglm(y,dist1,dist2,n,h,h.knn,kind.of.kernel,family,weights,
    rel.gvar,eff.rank,maxiter,eps1,eps2)
}

\details{These are not to be called by the user}
\keyword{internal}
