
\name{ldblm-inl}
\alias{ldblm-internal}
\alias{pred.train.sample}
\alias{kernel.number}
\alias{h.knn.funct}
\alias{controls_ldblm}

\title{Internal  functions}
\description{Internal  functions}
\usage{
pred.train.sample(y,dist1,dist2,n,h,h.knn,kind.of.kernel,weights,
        rel.gvar,eff.rank)
kernel.number(u,j=1)
h.knn.funct(dist,k=3)
controls_ldblm(dist1,dist2,user_h,method,h.range,noh,k.knn,kind.of.kernel,y,weights)
}

\details{These are not to be called by the user}
\keyword{internal}
