

\name{dbglm-inl}
\alias{dbglm-internal}
\alias{control_family}
\alias{controls_dbglm}

\title{Internal  functions}
\description{Internal  functions}
\usage{
control_family(family)
controls_dbglm(distance,weights,offset,rel.gvar,maxiter,eps1,eps2,y)
}

\details{Not to be called by users}
\keyword{internal}
